refactor ShowClassification into a component
