+++ static/js/components/Source.jsx
@@ -10,6 +10,8 @@ import Plot from './Plot';
 import CommentList from './CommentList';
 import ClassificationList from './ClassificationList';
 import ClassificationForm from './ClassificationForm';
+import ShowClassification from './ShowClassification';
+
 
 import ThumbnailList from './ThumbnailList';
 import SurveyLinkList from './SurveyLinkList';
@@ -70,67 +72,6 @@ const Source = ({ route }) => {
     );
   }
 
-  const groupBy = (array, key) => array.reduce((result, cv) => {
-    // if we've seen this key before, add the value, else generate
-    // a new list for this key
-    (result[cv[key]] = result[cv[key]] || []).push(cv);
-    return result;
-  }, {});
-  function showClassification() {
-    // Here we compute the most recent non-zero probability class for each taxonomy
-
-    const filteredClasses = source.classifications.filter((i) => i.probability > 0);
-    const groupedClasses = groupBy(filteredClasses, 'taxonomy_id');
-    const sortedClasses = [];
-
-    Object.keys(groupedClasses).forEach((item) => sortedClasses.push(
-      groupedClasses[item].sort((a, b) => (a.modified < b.modified ? 1 : -1))
-    ));
-
-
-    if (sortedClasses.length > 0) {
-      return (
-        <div>
-          <b>Classification: </b>
-          {sortedClasses.map((c) => {
-            let name = taxonomyList.filter((i) => i.id === c[0].taxonomy_id);
-            if (name.length > 0) {
-              name = name[0].name;
-            }
-            // generate the tooltip for this classification, with an informative
-            // hover over.
-            return (
-              <Tooltip
-                key={`${c[0].modified}tt`}
-                disableFocusListener
-                disableTouchListener
-                title={(
-                  <>
-                    P=
-                    {c[0].probability}
-                    {' '}
-                    (
-                    {name}
-                    )
-                    <br />
-                    <i>{c[0].author_name}</i>
-                  </>
-                      )}
-              >
-                <Button key={`${c[0].modified}tb`}>{c[0].classification}</Button>
-              </Tooltip>
-            );
-          })}
-        </div>
-      );
-    } else {
-      return (
-        <span />
-      );
-    }
-  }
-
-
   return (
     <div className={styles.source}>
 
@@ -141,7 +82,10 @@ const Source = ({ route }) => {
         </div>
 
         <br />
-        {showClassification()}
+          <ShowClassification
+            classifications={source.classifications}
+            taxonomyList={taxonomyList}
+          />
         <b>
           Position (J2000):
         </b>
